0.2.0-BETA

+++ services/bime/src/main/java/bime/dto/StockAlertResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -18,8 +19,8 @@ public class StockAlertResponseDTO {
     private UUID orgId;
     private UUID variantId;
     private UUID locationId;
-    private int threshold;
+    private BigDecimal threshold;
     @Schema(description = "On-hand quantity at the moment this alert was triggered")
-    private int quantity;
+    private BigDecimal quantity;
     private LocalDateTime triggeredAt;
 }
